pyproject.toml
[tool.poetry]
name = "correcao-limites-cafe"
version = "0.1.0"
description = ""
authors = ["André Carvalho <alac1984@gmail.com>"]
readme = "README.md"
packages = [{include = "correcao_limites_cafe"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.5.2"
pydantic = "^2.9.2"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.5.0"
black = "^24.8.0"
mypy = "^1.11.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
